Introduce a 'Shrink' method
Currently as the buffer is doubled in size when the number
of ActualWords is equal to the number required to hold the full bitmask -1
the size of the internal buffer of longs is doubled.

This can mean for certain use-case (large static BitMaps for example) that
once created do not change, it is desireable to reduce their overall memory
usage where possible.

I believe that setting the _Buffer length to that of _ActualWords is a safe
and effective mechanism for achieving this goal. (The downside of course is
that if the Shrink method is called, and then a bit is set, the new buffer length
may be greater than the pre-shrink, but another call to Shrink will free up this
excess once more)