Add a custom (simple) binary serializer
Adds a class that can be used to serialize (and deserialize)
instances of EwahCompressedBitArray to and from a pure
byte array.

Removes some of the 'bloat' associated with the default BinaryFormatters.

Minimal benefit for a small number of large BitArrays, but saves space when
serialising large numbers of small BitArrays.